requirements.txt
numpy
opencv-python
pillow
matplotlib
pytesseract
pyyaml
pandas
